# apps/docs/static/api/rest/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
    "/api/v1/experiments": {
      "post": {
        "description": "Creates a new experiment with the provided configuration",
        "summary": "Create a new experiment",
        "tags": ["experiments"],
        "parameters": [],
        "operationId": "experiments.createExperiment",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "The name of the experiment"
                  },
                  "description": {
                    "type": "string",
                    "description": "Optional description of the experiment"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "provisioning",
                      "provisioning_failed",
                      "active",
                      "stale",
                      "archived",
                      "published"
                    ],
                    "description": "Initial status of the experiment"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["private", "public"],
                    "description": "Experiment visibility setting"
                  },
                  "embargoIntervalDays": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": 0,
                    "description": "Embargo period in days"
                  },
                  "members": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "role": {
                          "type": "string",
                          "enum": ["admin", "member"]
                        }
                      },
                      "required": ["userId"]
                    },
                    "description": "Optional array of member objects with userId and role"
                  },
                  "protocols": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "protocolId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "order": {
                          "type": "integer"
                        }
                      },
                      "required": ["protocolId"]
                    },
                    "description": "Optional array of protocol objects with protocolId and order to associate with the experiment"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Returns a list of experiments based on the specified filter criteria, including search.",
        "summary": "List experiments",
        "tags": ["experiments"],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter experiments by relationship to the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": ["my", "member", "related"]
                }
              }
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter experiments by their status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "provisioning",
                    "provisioning_failed",
                    "active",
                    "stale",
                    "archived",
                    "published"
                  ]
                }
              }
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for experiment name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "operationId": "experiments.listExperiments",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "provisioning",
                          "provisioning_failed",
                          "active",
                          "stale",
                          "archived",
                          "published"
                        ]
                      },
                      "visibility": {
                        "type": "string",
                        "enum": ["private", "public"]
                      },
                      "embargoIntervalDays": {
                        "type": "integer"
                      },
                      "createdBy": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "columns": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                },
                                "type_name": {
                                  "type": "string"
                                },
                                "type_text": {
                                  "type": "string"
                                }
                              },
                              "required": ["name", "type_name", "type_text"]
                            }
                          },
                          "rows": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "string",
                                "nullable": true
                              }
                            }
                          },
                          "totalRows": {
                            "type": "integer"
                          },
                          "truncated": {
                            "type": "boolean"
                          }
                        },
                        "required": ["columns", "rows", "totalRows", "truncated"]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "description",
                      "status",
                      "visibility",
                      "embargoIntervalDays",
                      "createdBy",
                      "createdAt",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{id}": {
      "get": {
        "description": "Returns detailed information about a specific experiment",
        "summary": "Get experiment details",
        "tags": ["experiments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the experiment"
          }
        ],
        "operationId": "experiments.getExperiment",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "provisioning",
                        "provisioning_failed",
                        "active",
                        "stale",
                        "archived",
                        "published"
                      ]
                    },
                    "visibility": {
                      "type": "string",
                      "enum": ["private", "public"]
                    },
                    "embargoIntervalDays": {
                      "type": "integer"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "type_name": {
                                "type": "string"
                              },
                              "type_text": {
                                "type": "string"
                              }
                            },
                            "required": ["name", "type_name", "type_text"]
                          }
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string",
                              "nullable": true
                            }
                          }
                        },
                        "totalRows": {
                          "type": "integer"
                        },
                        "truncated": {
                          "type": "boolean"
                        }
                      },
                      "required": ["columns", "rows", "totalRows", "truncated"]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "status",
                    "visibility",
                    "embargoIntervalDays",
                    "createdBy",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "Updates an existing experiment with the provided changes",
        "summary": "Update experiment",
        "tags": ["experiments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the experiment"
          }
        ],
        "operationId": "experiments.updateExperiment",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "Updated experiment name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Updated experiment description"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "provisioning",
                      "provisioning_failed",
                      "active",
                      "stale",
                      "archived",
                      "published"
                    ],
                    "description": "Updated experiment status"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["private", "public"],
                    "description": "Updated visibility setting"
                  },
                  "embargoIntervalDays": {
                    "type": "integer",
                    "description": "Updated embargo period in days"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "provisioning",
                        "provisioning_failed",
                        "active",
                        "stale",
                        "archived",
                        "published"
                      ]
                    },
                    "visibility": {
                      "type": "string",
                      "enum": ["private", "public"]
                    },
                    "embargoIntervalDays": {
                      "type": "integer"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "type_name": {
                                "type": "string"
                              },
                              "type_text": {
                                "type": "string"
                              }
                            },
                            "required": ["name", "type_name", "type_text"]
                          }
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string",
                              "nullable": true
                            }
                          }
                        },
                        "totalRows": {
                          "type": "integer"
                        },
                        "truncated": {
                          "type": "boolean"
                        }
                      },
                      "required": ["columns", "rows", "totalRows", "truncated"]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "status",
                    "visibility",
                    "embargoIntervalDays",
                    "createdBy",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deletes an experiment and all associated data",
        "summary": "Delete experiment",
        "tags": ["experiments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the experiment"
          }
        ],
        "operationId": "experiments.deleteExperiment",
        "responses": {
          "204": {
            "description": "204"
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "OpenJII API",
    "version": "1.0.0",
    "description": "REST API documentation for the OpenJII platform"
  },
  "servers": [
    {
      "url": "http://localhost:3020",
      "description": "Local development"
    }
  ]
}